+++ CHANGELOG.rst
@@ -7,7 +7,7 @@ Changelog
 Features:
 
 - When reading CSV files, assume columns in same order as defined in class if not given in file.
-- Add the ``Integer``, ``Complex``, and ``Vector`` feature types.
+- Add the ``Integer``, ``Complex``, ``Vector``, and ``Tag`` feature types.
 - Add the ``Model.predict`` ``yield_inputs`` parameter.
 
 Bug fixes: